Add PostgreSQL LTREE column to store tree paths
This adds a column to microcosms that will be the store of the pre-calculated
tree of microcosms and allows very quick traversal of the tree and quick
querying of ancestors and children.

For example the following query would return the ancestors of microcosm id = 3
in order (oldest first):

SELECT m.*
  FROM microcosms m
      ,(SELECT microcosm_id
              ,path
          FROM microcosms
         WHERE microcosm_id = 3) c
 WHERE m.path @> c.path
   AND m.microcosm_id <> c.microcosm_id
 ORDER BY m.path ASC;

Requires that CREATE EXTENSION IF NOT EXISTS ltree; is run after the installation
of postgresql-contrib.