Added root subforums and fixed the permisions model
This change means that every site has only one microcosm that is the root
microcosm for the site. This parents all other microcosms, so that the home
page of the site is in fact a view of a microcosm.

This means that permissions can now be edited across the entire site, and
the moderators can be inherited all the way down the tree of microcosms.

Further, the permissions model will now treat the roles that belong to the
root microcosm as being the default roles.