Added API for returning a list of all microcosms on the site
This includes returning enough information to thread a tree together, and the
data is ordered by path to make this simple.

Additionally we return the level of the microcosm, which allows the list to be
easily rendered with indentation to illustrate the tree. This is useful when
rendering the list of forums in an HTML SELECT element whilst wishing to
communicate the tree structure.