Added more efficient sequencing of microcosms
This adds a sequence column to microcosms and then will calculate the
ordering of that across a site based on the sum of items and comments
that are within that and all child microcosms.

It also allows the possibility in the future to disable automatic ordering
and for the site to define their own order for microcosms.