Fix permissions error within search
This allowed you to discover the existence of a private forum if you
knew the name of it. The permissions were correctly enforced and would
prevent access, but the permissions were not correctly applied to the
search set and so would show you that the item existed.

Permissions are based on the premise that you can see things in a
forum for which you do have permission... but this is not true of
child forums where you do not have permission on those. This was
implemented correctly in microcosms.go:getChildMicrocosms() but not
implemented correctly in search_fulltext.go. I've made search_metadata.go
consistent with this new approach though that seems less likely to
be vulnerable (as you would need to know the forum identifier)