Modified the cleanup queries to overcome an issue where the queries never finished
The cardinality of the JOIN to cleanup items based on "delete all updates except the latest"
was using a few hash joins and these datasets were in the region of 3M rows producing 9T
combinations.

To overcome this I've moved from JOINs to an EXCEPT which doesn't attempt to iterate and match
but instead doe a hash compare rather than hash join.

Before this, the query would not complete in 24h, but now the query completes in about 3s.

As the query is more performant I have also made it run more often so that the Following
page performance is increased.